Add Basic Error Handlers and Verify Data

+++ backend/main.js
@@ -1,46 +1,62 @@
 const OBEX_utils = require('./obex_utils'),
 	  sleep = require('./sleep'),
 	  
-	  config = require('./config'),
-	  
 	  expandHomeDir = require('expand-home-dir'),
+	  fs = require('fs'),
 	  
-	  fs = require('fs');
+	  config = require('./config');
 
-function pollForNewData(macAddr){
-	const device = OBEX_utils.mount(macAddr, expandHomeDir(config.device_directory)),
-		  
-		  files = fs.readdirSync(
-			  expandHomeDir(config.receive_directory)
-		  );
-	
-	files.forEach(function(file){
-		const filePath = expandHomeDir(config.receive_directory) + file,
-			  
-			  newFile = fs.readFileSync(filePath, { encoding: "utf8" });
-		
-		// ENSURE INTEGRITY
-		
-		// Log File to MongoDB
-		//  - Poll via API
-		
-		fs.unlinkSync(filePath);
+function pollForNewData(macAddr) {
+	// Mount Device and Attempt to Read Directory for New Data
+	try {
+		const device = OBEX_utils.mount(macAddr, expandHomeDir(config.device_directory)),
+
+		files = fs.readdirSync(
+			expandHomeDir(config.receive_directory)
+		);
+	} catch (e) {
+		// TODO: DUMP ERRORS TO LOG
+	}
+
+	files.forEach(function (file) {
+		try {
+			const filePath = expandHomeDir(config.receive_directory) + file,
+				  
+				  newFile = fs.readFileSync(filePath, {
+					  encoding: "utf8"
+				  }),
+				  
+				  data = JSON.parse(newFile);
+
+			if (data.check != "9dcec4e5sd7f890s")
+				throw new Error("Error Parsing Data...");
+
+			// Log File to MongoDB
+
+			fs.unlinkSync(filePath);
+		} catch (e) {
+			// TODO: DUMP ERRORS TO LOG
+		}
 	});
-	
-	setTimeout(function(){
+
+	setTimeout(function () {
 		sleep(1);
-		
-		OBEX_utils.unmount(device);
-		
+
+		try {
+			OBEX_utils.unmount(device);
+		} catch (e) {
+			// TODO: DUMP ERRORS TO LOG
+		}
+
 		// Recursively Polls
 		pollForNewData(macAddr);
 	}, 5000);
 }
 
 const devices = [
-	process.argv[2];
+	process.argv[2]
 ];
 
-devices.forEach(function(device){
+devices.forEach(function (device) {
 	pollForNewData(device);
 });
